+++ src/Report/Xml/Facade.php
@@ -27,7 +27,6 @@ use function strlen;
 use function substr;
 use DateTimeImmutable;
 use DOMDocument;
-use SebastianBergmann\CodeCoverage\CodeCoverage;
 use SebastianBergmann\CodeCoverage\Driver\PathExistsButIsNotDirectoryException;
 use SebastianBergmann\CodeCoverage\Driver\WriteOperationFailedException;
 use SebastianBergmann\CodeCoverage\Node\AbstractNode;
@@ -50,9 +49,11 @@ final class Facade
     }
 
     /**
+     * @psalm-param array<string, array{size: string, status: string}> $tests
+     *
      * @throws XmlException
      */
-    public function process(CodeCoverage $coverage, string $target): void
+    public function process(DirectoryNode $report, array $tests, string $target): void
     {
         if (substr($target, -1, 1) !== DIRECTORY_SEPARATOR) {
             $target .= DIRECTORY_SEPARATOR;
@@ -61,14 +62,10 @@ final class Facade
         $this->target = $target;
         $this->initTargetDirectory($target);
 
-        $report = $coverage->getReport();
-
-        $this->project = new Project(
-            $coverage->getReport()->name()
-        );
+        $this->project = new Project($report->name());
 
         $this->setBuildInformation();
-        $this->processTests($coverage->getTests());
+        $this->processTests($tests);
         $this->processDirectory($report, $this->project);
 
         $this->saveDocument($this->project->asDom(), 'index');
@@ -215,6 +212,9 @@ final class Facade
         $functionObject->setTotals((string) $function['executableLines'], (string) $function['executedLines'], (string) $function['coverage']);
     }
 
+    /**
+     * @psalm-param array<string, array{size: string, status: string}> $tests
+     */
     private function processTests(array $tests): void
     {
         $testsObject = $this->project->tests();